@danmosemsft 2018-03-24
@stilettk thanks! Can you check we already have sufficient tests for this?