@stilettk 2018-03-24
> Comma after fails?

Will fix!

> Is performing these three reverses faster than a single explicit loop?

Can you please explain what algorithm with a single loop do you mean? Please consider that the lengths of swapped segments can be different, so simple swapping in a loop won't work.

> Can you check we already have sufficient tests for this?

Unfortunately, we don't. There are tests of `InsertRange` in CoreFX, but only inserting of `ICollection` is tested (see [#14876 (comment)](https://github.com/dotnet/coreclr/issues/14876#issuecomment-375137324)). This is because test collection returns `IEnumerable<T>`, which is still an array underneath (`collection is ICollection<T>` returns true):
```
public IEnumerable<T> ConstructTestCollection(T[] items)
{
    return items;
}
```

This needs to be fixed in another PR later, because tests are in another repo. Anyway, I have changed `ConstructTestCollection` method to return pure `IEnumerable<T>`, and tests have succeeded. 

If you mean performance tests, we don't have them for `InsertRange` either. Again, it needs to be done in separate issue since this is another repo.