@safern 2018-03-30
Note that there was some coverage added for this yesterday in: https://github.com/dotnet/corefx/pull/28465