@jkotas 2018-12-01
The earlier change that this is improving upon was found to be problematic and reverted in #20471.

Thanks for your PR anyway.